infra/requirements.txt
aws-cdk-lib>=2.70.0
constructs>=10.0.0
boto3>=1.28.0
